+++ src/index.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import mongodb from "mongodb";
-import Router from "./modules/Router.js";
+import RouteManager from "./modules/RouteManager.js";
 
 const uri = process.env.URI!;
 const port = process.env.PORT!;
@@ -11,7 +11,7 @@ const client = new mongodb.MongoClient(uri, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
-const router = new Router(app, client);
+const router = new RouteManager(app, client);
 
 console.log(`Connecting`);
 await client.connect();
